+++ persistent/Database/Persist.hs
@@ -77,7 +77,7 @@ infixr 3 =., +=., -=., *=., /=.
 -- Similar to `updateWhere` which is shown in the above example you can use other functions present in the module "Database.Persist.Class". Note that the first parameter of `updateWhere` is [`Filter` val] and second parameter is [`Update` val]. By comparing this with the type of `==.` and `=.`, you can see that they match up in the above usage.
 --
 -- The above query when applied on <#dataset dataset-1>, will produce this:
--- 
+--
 -- > +-----+-----+--------+
 -- > |id   |name |age     |
 -- > +-----+-----+--------+
@@ -195,7 +195,7 @@ infix 4 ==., <., <=., >., >=., !=.
 -- > |1    |SPJ  |40   |
 -- > +-----+-----+-----+
 
-f ==. a  = Filter f (Left a) Eq
+f ==. a  = Filter f (FilterValue a) Eq
 
 -- | Non-equality check.
 --
@@ -214,7 +214,7 @@ f ==. a  = Filter f (Left a) Eq
 -- > |2    |Simon|41   |
 -- > +-----+-----+-----+
 
-f !=. a = Filter f (Left a) Ne
+f !=. a = Filter f (FilterValue a) Ne
 
 -- | Less-than check.
 --
@@ -233,7 +233,7 @@ f !=. a = Filter f (Left a) Ne
 -- > |1    |SPJ  |40   |
 -- > +-----+-----+-----+
 
-f <. a  = Filter f (Left a) Lt
+f <. a  = Filter f (FilterValue a) Lt
 
 -- | Less-than or equal check.
 --
@@ -252,7 +252,7 @@ f <. a  = Filter f (Left a) Lt
 -- > |1    |SPJ  |40   |
 -- > +-----+-----+-----+
 
-f <=. a  = Filter f (Left a) Le
+f <=. a  = Filter f (FilterValue a) Le
 
 -- | Greater-than check.
 --
@@ -271,7 +271,7 @@ f <=. a  = Filter f (Left a) Le
 -- > |2    |Simon|41   |
 -- > +-----+-----+-----+
 
-f >. a  = Filter f (Left a) Gt
+f >. a  = Filter f (FilterValue a) Gt
 
 -- | Greater-than or equal check.
 --
@@ -290,7 +290,7 @@ f >. a  = Filter f (Left a) Gt
 -- > |2    |Simon|41   |
 -- > +-----+-----+-----+
 
-f >=. a  = Filter f (Left a) Ge
+f >=. a  = Filter f (FilterValue a) Ge
 
 infix 4 <-., /<-.
 (<-.), (/<-.) :: forall v typ.  PersistField typ => EntityField v typ -> [typ] -> Filter v
@@ -328,7 +328,7 @@ infix 4 <-., /<-.
 -- > |1    |SPJ  |40   |
 -- > +-----+-----+-----+
 
-f <-. a = Filter f (Right a) In
+f <-. a = Filter f (FilterValues a) In
 
 -- | Check if value is not in given list.
 --
@@ -347,7 +347,7 @@ f <-. a = Filter f (Right a) In
 -- > |2    |Simon|41   |
 -- > +-----+-----+-----+
 
-f /<-. a = Filter f (Right a) NotIn
+f /<-. a = Filter f (FilterValues a) NotIn
 
 infixl 3 ||.
 (||.) :: forall v. [Filter v] -> [Filter v] -> [Filter v]